refactor: currencies service pkg
