+++ app/common/currency.go
@@ -2,26 +2,36 @@ package common
 
 import (
 	"fmt"
-	"log/slog"
 
 	"github.com/google/wire"
 
 	"github.com/openmeterio/openmeter/openmeter/currencies"
-	currencyAdapter "github.com/openmeterio/openmeter/openmeter/currencies/adapter"
+	currencyadapter "github.com/openmeterio/openmeter/openmeter/currencies/adapter"
 	"github.com/openmeterio/openmeter/openmeter/currencies/service"
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 )
 
 var Currency = wire.NewSet(
+	NewCurrencyAdapter,
 	NewCurrencyService,
 )
 
-func NewCurrencyService(logger *slog.Logger, db *entdb.Client) (currencies.CurrencyService, error) {
-	adapter, err := currencyAdapter.New(currencyAdapter.Config{
+func NewCurrencyAdapter(db *entdb.Client) (currencies.Repository, error) {
+	repo, err := currencyadapter.New(currencyadapter.Config{
 		Client: db,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to create currency adapter: %w", err)
 	}
-	return service.New(adapter), nil
+
+	return repo, nil
+}
+
+func NewCurrencyService(repo currencies.Repository) (currencies.Service, error) {
+	s, err := service.New(repo)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create currency service: %w", err)
+	}
+
+	return s, nil
 }
