+++ openmeter/currencies/repository.go
@@ -7,11 +7,19 @@ import (
 	"github.com/openmeterio/openmeter/pkg/pagination"
 )
 
-type Adapter interface {
+type Repository interface {
 	entutils.TxCreator
 
+	CurrencyRepository
+	CostBasisRepository
+}
+
+type CurrencyRepository interface {
 	ListCustomCurrencies(ctx context.Context, params ListCurrenciesInput) (pagination.Result[Currency], error)
 	CreateCurrency(ctx context.Context, params CreateCurrencyInput) (Currency, error)
+}
+
+type CostBasisRepository interface {
 	CreateCostBasis(ctx context.Context, params CreateCostBasisInput) (CostBasis, error)
 	ListCostBases(ctx context.Context, params ListCostBasesInput) (pagination.Result[CostBasis], error)
 }
